requirements.txt
flask==2.3.2
requests==2.31.0
pyjwt==2.8.0